feat: modal en proyectos
